(BSR) cultural survey: move api calls in api folder and in separate files
