+++ src/features/culturalSurvey/api/__mocks__/useCulturalSurveyAnswers.ts
@@ -1,0 +1,9 @@
+interface Props {
+  onSuccess: () => void
+  onError: (error: unknown) => void
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const useCulturalSurveyAnswersMutation = jest.fn(({ onSuccess, onError }: Props) => {
+  return { mutate: onSuccess }
+})
