+++ src/features/culturalSurvey/api/__mocks__/useCulturalSurveyQuestions.ts
@@ -1,8 +1,8 @@
 import { UseQueryResult } from 'react-query'
 
 import { CulturalSurveyQuestionsResponse } from 'api/gen'
+import { useCulturalSurveyQuestions as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/api/useCulturalSurveyQuestions'
 import { mockedCulturalSurveyQuestions } from 'features/culturalSurvey/fixtures/mockedCulturalSurveyQuestions'
-import { useCulturalSurveyQuestions as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/useCulturalSurvey'
 
 export const useCulturalSurveyQuestions: typeof actualUseCulturalSurveyQuestions = jest.fn(
   () =>
@@ -11,13 +11,3 @@ export const useCulturalSurveyQuestions: typeof actualUseCulturalSurveyQuestions
       unknown
     >)
 )
-
-interface Props {
-  onSuccess: () => void
-  onError: (error: unknown) => void
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const useCulturalSurveyAnswersMutation = jest.fn(({ onSuccess, onError }: Props) => {
-  return { mutate: onSuccess }
-})
