+++ src/features/culturalSurvey/pages/CulturalSurveyQuestions.tsx
@@ -11,6 +11,8 @@ import {
   CulturalSurveyQuestion,
   CulturalSurveyQuestionEnum,
 } from 'api/gen'
+import { useCulturalSurveyAnswersMutation } from 'features/culturalSurvey/api/useCulturalSurveyAnswers'
+import { useCulturalSurveyQuestions } from 'features/culturalSurvey/api/useCulturalSurveyQuestions'
 import { CulturalSurveyCheckbox } from 'features/culturalSurvey/components/CulturalSurveyCheckbox'
 import { CulturalSurveyPageHeader } from 'features/culturalSurvey/components/layout/CulturalSurveyPageHeader'
 import { useCulturalSurveyContext } from 'features/culturalSurvey/context/CulturalSurveyContextProvider'
@@ -19,10 +21,6 @@ import {
   mapQuestionIdToPageTitle,
   removeSubQuestionsToDisplay,
 } from 'features/culturalSurvey/helpers/utils'
-import {
-  useCulturalSurveyQuestions,
-  useCulturalSurveyAnswersMutation,
-} from 'features/culturalSurvey/useCulturalSurvey'
 import { useCulturalSurveyProgress } from 'features/culturalSurvey/useCulturalSurveyProgress'
 import { useGetNextQuestion } from 'features/culturalSurvey/useGetNextQuestion'
 import { navigateToHome } from 'features/navigation/helpers'
